Fix to DbUtils.py729:738 for production publish
This section was breaking the production/test Db. The contents of `object_naming_info` are modifies somewhere else before here so that this IF/ELSE is not needed and causes a break in the code. For now I commented it out incase it was needed for something else.

Changes to be committed:
	modified:   bco_api/api/scripts/utilities/DbUtils.py